+++ CHANGELOG.rst
@@ -22,9 +22,8 @@ Each individual change should have a link to the pull request after the descript
 - chore: refactored unit weight handling into method of WeightColumnMixin, tightened up handling
 - feat: added AggregateColumnsOverRowTransformer `#385 <https://github.com/azukds/tubular/issues/385>_`
 - chore: renamed AggregateRowOverColumnsTransformer to AggregateRowsOverColumnTransformer
-- placeholder
-- placeholder
-
+- feat: optimisation changes to fit methods in imputers file
+- feat: optimisation changes to transform method of NullIndicator
 
 1.4.8 (03/09/25)
 ----------------